requirements.txt
Flask==2.3.2
pydantic==1.10.9
google-cloud-firestore==2.11.1
